Moved overlays to dedicated package
