+++ src/main/java/shortestpath/overlay/PathTileOverlay.java
@@ -1,4 +1,4 @@
-package shortestpath;
+package shortestpath.overlay;
 
 import com.google.inject.Inject;
 
@@ -23,6 +23,11 @@ import net.runelite.api.coords.LocalPoint;
 import net.runelite.client.ui.overlay.Overlay;
 import net.runelite.client.ui.overlay.OverlayLayer;
 import net.runelite.client.ui.overlay.OverlayPosition;
+import shortestpath.PrimitiveIntList;
+import shortestpath.ShortestPathPlugin;
+import shortestpath.TileCounter;
+import shortestpath.TileStyle;
+import shortestpath.WorldPointUtil;
 import shortestpath.pathfinder.CollisionMap;
 import shortestpath.pathfinder.PathStep;
 import shortestpath.transport.BankPickupRequirements;
